+++ examples/simulator/mechanics-simulator-leftventricle_pr.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2023 - 2025 ANSYS, Inc. and/or its affiliates.
+# Copyright (C) 2023 - 2026 ANSYS, Inc. and/or its affiliates.
 # SPDX-License-Identifier: MIT
 #
 #
@@ -74,7 +74,8 @@ import pyvista as pv
 from ansys.health.heart.examples import get_input_leftventricle
 import ansys.health.heart.models as models
 from ansys.health.heart.post.dpf_utils import ICVoutReader
-from ansys.health.heart.settings.material.material import ACTIVE, ANISO, ISO, Mat295
+from ansys.health.heart.settings.material.material import ACTIVE, ANISO, ISO, HGOFiber, Mat295
+from ansys.health.heart.settings.settings import FibersDRBM
 from ansys.health.heart.simulator import DynaSettings, MechanicsSimulator
 
 ###############################################################################
@@ -151,11 +152,28 @@ simulator.settings.load_defaults()
 # Compute fiber orientation
 # ~~~~~~~~~~~~~~~~~~~~~~~~~
 # Compute fiber orientation and plot the fibers.
-simulator.compute_fibers(method="D-RBM")
+
+# Use default rotation angles for the D-RBM method.
+fiber_settings = FibersDRBM()
+
+# Print settings
+print(fiber_settings)
+
+simulator.compute_fibers(fiber_settings=fiber_settings)
 
 # Plot the fiber orientation by streamlines.
 simulator.model.plot_fibers(n_seed_points=2000)
 
+###############################################################################
+# .. note::
+#    The D-RBM method is based on Doste et al. (2021) and in this implementation
+#    the rotation of the local coordinate system follows the right hand rule, where
+#    the transmural direction is pointing from endocardium to epicardium, the
+#    longitudinal direction is pointing from apex to base. Consequently, the
+#    circumferential direction is orthogonal to both and follows from the cross
+#    product of longtiduinal and transmural directions.
+
+###############################################################################
 
 ###############################################################################
 # Assign a material to the left ventricle.
@@ -167,7 +185,7 @@ myocardium = Mat295(
     iso=ISO(itype=-3, beta=2, kappa=1.0, k1=0.20e-3, k2=6.55),
     aniso=ANISO(
         atype=-1,
-        fibers=[ANISO.HGOFiber(k1=0.00305, k2=29.05), ANISO.HGOFiber(k1=1.25e-3, k2=36.65)],
+        fibers=[HGOFiber(k1=0.00305, k2=29.05), HGOFiber(k1=1.25e-3, k2=36.65)],
         k1fs=0.15e-3,
         k2fs=6.28,
     ),
@@ -180,6 +198,10 @@ simulator.model.left_ventricle.meca_material = myocardium
 ################################################################################
 # Compute the stress-free configuration
 # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+# Create a stiff region at the base of the ventricle
+simulator.compute_uhc()
+stiff_part = simulator.model.create_stiff_ventricle_base(threshold_left_ventricle=0.9)
 
 # Compute the stress-free configuration and estimate initial stress.
 report, stress_free_coord, guess_ed_coord = simulator.compute_stress_free_configuration()