+++ src/ansys/health/heart/exceptions.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2023 - 2025 ANSYS, Inc. and/or its affiliates.
+# Copyright (C) 2023 - 2026 ANSYS, Inc. and/or its affiliates.
 # SPDX-License-Identifier: MIT
 #
 #
@@ -21,6 +21,8 @@
 # SOFTWARE.
 
 """Custom exceptions for PyAnsys Heart."""
+
+from typing import Literal
 
 
 class LSDYNATerminationError(BaseException):
@@ -76,3 +78,14 @@ class WSLNotFoundError(FileNotFoundError):
 
 class MissingEnvironmentVariableError(EnvironmentError):
     """Exception raised when a required environment variable is missing."""
+
+
+class MissingMaterialError(ValueError):
+    """Exception raised when a required material is missing in the model."""
+
+    def __init__(self, part_name: str, material_type: Literal["EP", "Mechanical"]):
+        super().__init__(f"Part {part_name} has no {material_type} material assigned.")
+
+
+class PartAlreadyExistsError(ValueError):
+    """Exception raised when a part already exists."""