+++ src/ansys/health/heart/objects.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2023 - 2025 ANSYS, Inc. and/or its affiliates.
+# Copyright (C) 2023 - 2026 ANSYS, Inc. and/or its affiliates.
 # SPDX-License-Identifier: MIT
 #
 #
@@ -88,7 +88,7 @@ def _get_fill_data(
 
 
 def _get_global_cell_ids(mesh: pv.UnstructuredGrid, celltype: pv.CellType) -> np.ndarray:
-    """Get the global cell iID of a given cell type.
+    """Get the global cell IDs of a given cell type.
 
     Parameters
     ----------
@@ -227,8 +227,14 @@ class SurfaceMesh(pv.PolyData):
     @property
     def triangles(self):
         """Triangular faces of the surface ``num_faces`` x 3."""
-        faces = np.reshape(self.faces, (self.n_cells, 3 + 1))[:, 1:]
-        return faces
+        if self.n_cells == 0:
+            return np.empty((0, 3), dtype=int)
+
+        if self.is_all_triangles:
+            return self.regular_faces
+
+        else:
+            raise ValueError("Not all faces in the mesh are triangles.")
 
     @triangles.setter
     def triangles(self, value: np.ndarray):
@@ -723,7 +729,7 @@ class Mesh(pv.UnstructuredGrid):
                     continue
                 mesh.point_data[name] = fill_data
 
-        merged = pv.merge((self, mesh), merge_points=merge_points, main_has_priority=False)
+        merged = pv.merge((self, mesh), merge_points=merge_points)
         super().__init__(merged)
         return self
 
@@ -1043,7 +1049,7 @@ class Mesh(pv.UnstructuredGrid):
             PolyData representation of the lines to add.
         id : int
             ID of the surface to add. This ID is tracked as ``_line-id``.
-        name : str, optional
+        name : str, default: None
             Name of the added lines. The added lines are not tracked by default.
         """
         if not id:
@@ -1086,12 +1092,16 @@ class Mesh(pv.UnstructuredGrid):
         """
         if sid in list(self._surface_id_to_name.keys()):
             return SurfaceMesh(
-                self._get_submesh(sid, scalar="_surface-id").extract_surface(),
+                self._get_submesh(sid, scalar="_surface-id").extract_surface(
+                    algorithm="dataset_surface"
+                ),
                 name=self._surface_id_to_name[sid],
                 id=sid,
             )
         else:
-            return self._get_submesh(sid, scalar="_surface-id").extract_surface()
+            return self._get_submesh(sid, scalar="_surface-id").extract_surface(
+                algorithm="dataset_surface"
+            )
 
     def get_surface_by_name(self, name: str) -> Union[pv.PolyData, SurfaceMesh]:
         # ?: Return SurfaceMesh instead of PolyData?
@@ -1104,7 +1114,9 @@ class Mesh(pv.UnstructuredGrid):
 
     def get_lines(self, sid: int) -> pv.PolyData:
         """Get lines as a PyVista polydata object."""
-        return self._get_submesh(sid, scalar="_line-id").extract_surface()
+        return self._get_submesh(sid, scalar="_line-id").extract_surface(
+            algorithm="dataset_surface"
+        )
 
     def get_lines_by_name(self, name: str) -> pv.PolyData:
         """Get the lines associated with a given name."""