+++ src/ansys/health/heart/writer/custom_keywords/keywords/em_ep_isoch.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2023 - 2025 ANSYS, Inc. and/or its affiliates.
+# Copyright (C) 2023 - 2026 ANSYS, Inc. and/or its affiliates.
 # SPDX-License-Identifier: MIT
 #
 #
@@ -27,7 +27,7 @@ from ansys.dyna.core.lib.keyword_base import KeywordBase
 
 
 class EmEpIsoch(KeywordBase):
-    """LS-DYNA EM_EP_ISOCH keyword"""
+    """LS-DYNA EM_EP_ISOCH keyword."""
 
     keyword = "EM"
     subkeyword = "EP_ISOCH"