+++ src/ansys/health/heart/writer/ep_mechanics_writer.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2023 - 2025 ANSYS, Inc. and/or its affiliates.
+# Copyright (C) 2023 - 2026 ANSYS, Inc. and/or its affiliates.
 # SPDX-License-Identifier: MIT
 #
 #
@@ -24,7 +24,7 @@
 from typing import Callable, Optional
 
 from ansys.dyna.core.keywords import keywords
-from ansys.health.heart.models import FourChamber, HeartModel
+from ansys.health.heart.models import HeartModel
 from ansys.health.heart.settings.settings import SimulationSettings
 from ansys.health.heart.writer.base_writer import BaseDynaWriter
 from ansys.health.heart.writer.ep_writer import ElectrophysiologyDynaWriter
@@ -40,9 +40,6 @@ class ElectroMechanicsDynaWriter(MechanicsDynaWriter, ElectrophysiologyDynaWrite
         model: HeartModel,
         settings: Optional[SimulationSettings] = None,
     ) -> None:
-        if isinstance(model, FourChamber):
-            model._create_atrioventricular_isolation()
-
         BaseDynaWriter.__init__(self, model=model, settings=settings)
 
         self.kw_database = ElectroMechanicsDecks()
@@ -67,12 +64,6 @@ class ElectroMechanicsDynaWriter(MechanicsDynaWriter, ElectrophysiologyDynaWrite
         -----
         You do not need to write mesh files if a Dynain file is given.
         """
-        if isinstance(self.model, FourChamber):
-            self.model.left_atrium.fiber = True
-            self.model.left_atrium.active = True
-            self.model.right_atrium.fiber = True
-            self.model.right_atrium.active = True
-
         MechanicsDynaWriter.update(self, dynain_name=dynain_name, robin_bcs=robin_bcs)
 
         if self.model.conduction_mesh.number_of_cells != 0:
@@ -88,6 +79,10 @@ class ElectroMechanicsDynaWriter(MechanicsDynaWriter, ElectrophysiologyDynaWrite
 
         self._update_ep_settings(beam_pid)
         self._update_stimulation()
+
+        # Add ECG Output
+        if hasattr(self.model, "electrodes") and len(self.model.electrodes) != 0:
+            self._update_ECG_coordinates()
 
         # coupling parameters
         coupling_str = (