+++ tests/heart/post/test_postprocess.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2023 - 2025 ANSYS, Inc. and/or its affiliates.
+# Copyright (C) 2023 - 2026 ANSYS, Inc. and/or its affiliates.
 # SPDX-License-Identifier: MIT
 #
 #
@@ -49,11 +49,12 @@ def get_left_ventricle():
 
 
 @pytest.mark.requires_dpf
-def test_compute_thickness(get_left_ventricle):
+def test_compute_thickness(get_left_ventricle, monkeypatch):
     test_dir = get_left_ventricle[0]
     model = get_left_ventricle[1]
     d3plot = os.path.join(os.path.join(test_dir, "main", "d3plot"))
 
+    monkeypatch.setenv("ANSYS_DPF_ACCEPT_LA", "Y")
     s = AhaStrainCalculator(model, d3plot)
     lines = s._compute_thickness_lines()
     assert len(lines) == 2
@@ -76,7 +77,8 @@ def test_compute_strain(get_left_ventricle):
 
 
 @pytest.mark.requires_dpf
-def test_compute_myocardial_strain(get_left_ventricle):
+def test_compute_myocardial_strain(get_left_ventricle, monkeypatch):
+    monkeypatch.setenv("ANSYS_DPF_ACCEPT_LA", "Y")
     test_dir = get_left_ventricle[0]
     model = get_left_ventricle[1]
     d3plot = os.path.join(os.path.join(test_dir, "main", "d3plot"))
@@ -87,7 +89,8 @@ def test_compute_myocardial_strain(get_left_ventricle):
 
 
 @pytest.mark.requires_dpf
-def test_compute_aha_strain(get_left_ventricle):
+def test_compute_aha_strain(get_left_ventricle, monkeypatch):
+    monkeypatch.setenv("ANSYS_DPF_ACCEPT_LA", "Y")
     test_dir = get_left_ventricle[0]
     model = get_left_ventricle[1]
     d3plot = os.path.join(os.path.join(test_dir, "main", "d3plot"))
@@ -99,8 +102,9 @@ def test_compute_aha_strain(get_left_ventricle):
 
 
 @pytest.mark.requires_dpf
-def test_plot_aha_bullseye():
+def test_plot_aha_bullseye(monkeypatch):
     """Test plotting AHA bullseye plot."""
+    monkeypatch.setenv("ANSYS_DPF_ACCEPT_LA", "Y")
 
     # Create the fake data
     data = np.arange(17) + 1
@@ -118,7 +122,8 @@ def test_plot_aha_bullseye():
 
 
 @pytest.mark.requires_dpf
-def test_zerop_post(get_left_ventricle):
+def test_zerop_post(get_left_ventricle, monkeypatch):
+    monkeypatch.setenv("ANSYS_DPF_ACCEPT_LA", "Y")
     test_dir = get_left_ventricle[0]
     model = get_left_ventricle[1]
     dct = zerop_post(os.path.join(test_dir, "zerop"), model)
@@ -137,11 +142,13 @@ def test_zerop_post(get_left_ventricle):
 @pytest.mark.requires_dpf
 class TestSystemModelPost:
     @pytest.fixture
-    def system_model(self, get_left_ventricle):
+    def system_model(self, get_left_ventricle, monkeypatch):
+        monkeypatch.setenv("ANSYS_DPF_ACCEPT_LA", "Y")
         test_dir = get_left_ventricle[0]
         return SystemModelPost(os.path.join(test_dir, "main"))
 
-    def test_plot_pv_loop(self, system_model):
+    def test_plot_pv_loop(self, system_model, monkeypatch):
+        monkeypatch.setenv("ANSYS_DPF_ACCEPT_LA", "Y")
         ef = system_model.get_ejection_fraction()
         fig = system_model.plot_pv_loop(ef=ef)
         fig.savefig("pv_{0:d}.png".format(0))