+++ tests/heart/pre/test_conduction_path.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2023 - 2025 ANSYS, Inc. and/or its affiliates.
+# Copyright (C) 2023 - 2026 ANSYS, Inc. and/or its affiliates.
 # SPDX-License-Identifier: MIT
 #
 #
@@ -36,8 +36,18 @@ from ansys.health.heart.pre.conduction_path import (
     _create_path_on_surface_center,
     _path_merge,
 )
-from ansys.health.heart.settings.material.ep_material import EPMaterial
 from tests.heart.conftest import get_assets_folder, get_fourchamber, get_fullheart
+
+
+def test_define_12lead_electrodes():
+    model = get_fourchamber()
+    _ = model.define_12lead_electrodes()
+
+    assert len(model.electrodes) == 10
+    assert model.electrodes[0].name == "V1"
+    assert np.allclose(
+        model.electrodes[0].xyz, np.array([[-16.07677295, 9.77005318, 364.10423886]]), atol=1e-3
+    )
 
 
 def test_create_conductionbeams_on_surface():
@@ -185,7 +195,7 @@ def test_conductionbeams_from_k():
         merge_apex=False,
     )
     assert l_pj.name == ConductionPathType.LEFT_PURKINJE
-    assert l_pj.ep_material == EPMaterial.DummyMaterial()
+    assert l_pj.ep_material is None
     ref0 = pv.read(os.path.join(folder, "left_purkinje.vtp"))
     assert meshes_equal(l_pj.mesh, ref0)
     assert l_pj.is_connected.sum() == 1114
@@ -243,10 +253,6 @@ def simple_conduction_path():
     mesh = pv.PolyData(line_points, lines=line_lines)
     is_connected = np.array([0, 0])
 
-    # Dummy EPMaterial
-    class DummyMaterial:
-        pass
-
     # Minimal ConductionPathType
     class DummyType:
         pass
@@ -258,7 +264,7 @@ def simple_conduction_path():
         id=1,
         is_connected=is_connected,
         relying_surface=surface,
-        material=DummyMaterial(),
+        material=None,
     )
     return cp